+++ familyhelper-ui-view/packages/webapp-mobile/src/views/vim/layout/Layout.vue
@@ -1,20 +1,16 @@
 <script lang="ts">
-import vim from '@/vim'
-
-import { type VisualizerStore } from '@/store/modules/visualizer.ts'
-
 import { h } from 'vue'
 
 import VimMask from '@/components/vim/vimMask/VimMask.vue'
 
+import { ready } from '@/util/store.ts'
 import { getVisualizer } from '@/util/visualizer.ts'
 
 export default {
   name: 'LayoutComponent',
   setup() {
-    const visualizerStore = vim.ctx().store().vueStore<'visualizer', VisualizerStore>('visualizer')
     return () => {
-      if (!visualizerStore.ready) {
+      if (!ready.value) {
         return h(VimMask)
       }
       return h(getVisualizer().render('layout', h))